few comments

+++ stuff/Consumable.h
@@ -13,13 +13,16 @@ namespace wumpus_game {
     class BaseUnit;
     class Consumable :public Item {
     private:
-        const int hp_mod_;
-        const int mana_mod_;
+        const double hp_mod_;
+        const double mana_mod_;
     public:
         Consumable(const std::string &item_name, const double &weight, const double &volume, const double &delta_hp,
                    const double &delta_mana);
-        int get_delta_hp(){return hp_mod_;}
-        int get_delta_mana(){return mana_mod_;}
+
+        double get_delta_hp() { return hp_mod_; }
+
+        //prob not used
+        double get_delta_mana() { return mana_mod_; }
 
         void Consume(BaseUnit *);
